+++ __tests__/useLayoutEffect.test.js
@@ -1,8 +1,0 @@
-import { useLayoutEffect } from 'react'
-import { useIsomorphicLayoutEffect } from '../src/hooks'
-
-describe('when the environment is browser', () => {
-  it('should use useLayoutEffect', () => {
-    expect(useIsomorphicLayoutEffect).toBe(useLayoutEffect)
-  })
-})